Moving pages ordering
Put older landing page to be "About" then add templete Landing page with background image; added into router&navBar
