+++ src/App.js
@@ -1,27 +1,29 @@
 import React from "react";
 import Amplify from "aws-amplify";
 import { Switch, Route } from "react-router-dom";
-
-import AppWithAuth from "./components/AppWithAuth";
 import Landing from "./components/Landing";
+import AppWithAuth from "./components/AppWithAuth";
+import About from "./components/About";
 // import Login from "./components/Login";
 import Footer from "./components/Footer";
 import NavigationBar from "./components/NavigationBar";
 import SignUp from "./components/SignUp";
 import awsconfig from "./aws-exports";
-import "bootstrap/dist/css/bootstrap.min.css";
+// import "bootstrap/dist/css/bootstrap.min.css";
 
 Amplify.configure(awsconfig);
 
 export default function App() {
   return (
-    <div className="App">
+    <div>
       <NavigationBar />
       <Switch>
         <Route exact path="/">
           <Landing />
         </Route>
-        <Route path="/about">hi</Route>
+        <Route exact path="/about">
+          <About />
+        </Route>
         <Route path="/signup">
           <SignUp />
         </Route>
