+++ src/components/NavigationBar.js
@@ -1,23 +1,54 @@
-import React from 'react'
-import { Nav, Navbar  } from 'react-bootstrap'
+import React from "react";
+import { Nav, Navbar } from "react-bootstrap";
 import { Link, useLocation } from "react-router-dom";
 
 export default function NavigationBar() {
-    let location = useLocation();
-    return (
-        <Navbar bg="light" expand="lg" fixed="top">
-            <Navbar.Brand href="/">Community</Navbar.Brand>
-            <Nav>
-            <Nav.Item>
-                <Nav.Link componentClass={Link} href="/about" to="/about" active={location.pathname === '/about'}>About</Nav.Link>
-            </Nav.Item>
-            <Nav.Item>
-                <Nav.Link componentClass={Link} href="/login" to="/login" active={location.pathname === '/login'}>Login</Nav.Link>
-            </Nav.Item>
-            <Nav.Item>
-                <Nav.Link componentClass={Link} href="signup" to="/signup" active={location.pathname === '/signup'}>Signup</Nav.Link>
-            </Nav.Item>
-            </Nav>
-      </Navbar>
-    );
-};
+  let location = useLocation();
+  return (
+    <Navbar bg="light" expand="lg" fixed="top">
+      <Navbar.Brand href="/">Community</Navbar.Brand>
+      <Nav>
+        <Nav.Item>
+          <Nav.Link
+            componentClass={Link}
+            href="/"
+            to="/"
+            active={location.pathname === "/"}
+          >
+            Home
+          </Nav.Link>
+        </Nav.Item>
+        <Nav.Item>
+          <Nav.Link
+            componentClass={Link}
+            href="/about"
+            to="/about"
+            active={location.pathname === "/about"}
+          >
+            About
+          </Nav.Link>
+        </Nav.Item>
+        <Nav.Item>
+          <Nav.Link
+            componentClass={Link}
+            href="/login"
+            to="/login"
+            active={location.pathname === "/login"}
+          >
+            Login
+          </Nav.Link>
+        </Nav.Item>
+        <Nav.Item>
+          <Nav.Link
+            componentClass={Link}
+            href="signup"
+            to="/signup"
+            active={location.pathname === "/signup"}
+          >
+            Signup
+          </Nav.Link>
+        </Nav.Item>
+      </Nav>
+    </Navbar>
+  );
+}
